add movie detail

+++ domain/movie.go
@@ -3,7 +3,7 @@ package domain
 type Movie struct {
 	Id       int
 	Title    string
-	Rate     float32
+	Rate     float64
 	Abstruct string
 	FLink    string
 	Reviews  []string
